requirements.txt
psycopg2-binary
schedule
python-dotenv